% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{categorized_data}
\alias{categorized_data}
\title{An entry in the returned array is the index of the bin of the
linearly-binned raw continuous data.}
\usage{
categorized_data(raw, n_bins)
}
\arguments{
\item{raw}{matrix of raw continuous data.}

\item{n_bins}{A universal number of bins for all the variables.}
}
\description{
An entry in the returned array is the index of the bin of the
linearly-binned raw continuous data.
}
