% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{conditional_entropy}
\alias{conditional_entropy}
\title{Conditional entropy of variables in the data conditioned on
a given set of variables.}
\usage{
conditional_entropy(data, given)
}
\arguments{
\item{data}{matrix of data with variables of interests as columns and
observations as rows.}

\item{given}{matrix of data with the conditioned variables as columns and
observations as rows.}
}
\description{
Conditional entropy of variables in the data conditioned on
a given set of variables.
}
