% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-spanning-tree.R
\name{construct_correlation_spanning}
\alias{construct_correlation_spanning}
\title{Creates a minimum spanning tree connecting the sensors.}
\usage{
construct_correlation_spanning(inputArray, distance, ...)
}
\description{
Creates a minimum spanning tree connecting the sensors.
}
