% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{construct_random}
\alias{construct_random}
\title{Reconstruct a network from a random matrix, not taking time series into account.}
\usage{
construct_random(inputArray, inputThreshold, ...)
}
\arguments{
\item{inputArray}{inputArray}

\item{inputThreshold}{inputThreshold}

\item{...}{...}
}
\description{
Reconstruct a network from a random matrix, not taking time series into account.
}
