% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergent_cross_mapping.R
\name{convergent_cross_mapping_fit}
\alias{convergent_cross_mapping_fit}
\title{Graph reconstruction algorithm from time series data based on}
\usage{
convergent_cross_mapping_fit(
  TS,
  tau = 1,
  threshold_type = "range",
  cutoffs = list(c(0.95, Inf)),
  ...
)
}
\arguments{
\item{TS}{input matrix, N * L matrix consisting of L observations
from N sensors.}

\item{tau}{Number of time steps for a single time-lag.}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{cutoffs}{cutoffs for threshold function}

\item{...}{arguments}
}
\description{
Graph reconstruction algorithm from time series data based on
}
