% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix_fit}
\alias{correlation_matrix_fit}
\title{Reconstruction of graphs using the correlation matrix.}
\usage{
correlation_matrix_fit(TS, num_eigs = NULL, threshold_type = "range", ...)
}
\arguments{
\item{TS}{Matrix consisting of L observations from N sensors}

\item{num_eigs}{The number of eigenvalues to use. (This corresponds to the
amount of regularization.) The number of eigenvalues used must
be less than N.}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{...}{ARGUMENTS}
}
\description{
Reconstruction of graphs using the correlation matrix.
}
