% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create_graph}
\alias{create_graph}
\title{Utilities for creating and interacting with graph objects.}
\usage{
create_graph(A, mode = NULL, remove_self_loops = T)
}
\arguments{
\item{A}{input MATRIX}

\item{mode}{Character scalar, specifies how igraph should interpret the supplied matrix.}

\item{remove_self_loops}{If True, remove the diagonal of the matrix before creating the
graph object.}
}
\description{
Utilities for creating and interacting with graph objects.
}
