% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_energy_minimization.R
\name{free_energy_minimization_fit}
\alias{free_energy_minimization_fit}
\title{Reconstruction of graphs by minimizing a free energy of your data}
\usage{
free_energy_minimization_fit(TS, threshold_type = "degree", ...)
}
\arguments{
\item{TS}{Matrix consisting of L observations from N different variables.}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{...}{Arguments}
}
\description{
Reconstruction of graphs by minimizing a free energy of your data
}
