% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granger_causality.R
\name{granger_causality_fit}
\alias{granger_causality_fit}
\title{Graph reconstruction algorithm}
\usage{
granger_causality_fit(TS, lag = 1, threshold_type = "range", ...)
}
\arguments{
\item{TS}{Array consisting of L observations from N variables.}

\item{lag}{Time lag to consider.}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{...}{Arguments}
}
\description{
Graph reconstruction algorithm
}
