% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_lasso.R
\name{graphical_lasso_fit}
\alias{graphical_lasso_fit}
\title{Graph reconstruction algorithm.}
\usage{
graphical_lasso_fit(
  TS,
  alpha = 0.01,
  max_iter = 100,
  tol = 1e-04,
  threshold_type = "degree",
  ...
)
}
\arguments{
\item{TS}{Matrix consisting of L observations from N variables.}

\item{alpha}{Coefficient of penalization, higher values means more sparseness}

\item{max_iter}{Maximum number of iterations.}

\item{tol}{Stop the algorithm when the duality gap is below a certain threshold.}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{...}{Arguments}
}
\description{
Graph reconstruction algorithm.
}
