% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marchenko_pastur.R
\name{marchenko_pastur_fit}
\alias{marchenko_pastur_fit}
\title{Graph reconstruction algorithm.}
\usage{
marchenko_pastur_fit(
  TS,
  remove_largest = F,
  metric_distance = F,
  tol = 1e-15,
  threshold_type = "range",
  ...
)
}
\arguments{
\item{TS}{N * L matrix consisting of L observations from N variables.}

\item{remove_largest}{If False, all the eigenvectors associated to the
significant eigenvalues will be used to reconstruct the
de-noised empirical correlation matrix. If True, the
eigenvector associated to the largest eigenvalue is going to be excluded from
the recontruction step.}

\item{metric_distance}{If False, a signed graph is obtained. If True, the correlation
is transformed by defining a metric distance between each pair of nodes}

\item{tol}{avoid minor difference between python and r in case of calculating eigen values}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{...}{Arguments}
}
\description{
Graph reconstruction algorithm.
}
