% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{maximum_likelihood_estimation_fit}
\alias{maximum_likelihood_estimation_fit}
\title{Reconstruction of graphs using maximum likelihood estimation}
\usage{
maximum_likelihood_estimation_fit(
  TS,
  rate = 1,
  stop_criterion = T,
  threshold_type = "degree",
  ...
)
}
\arguments{
\item{TS}{Matrix consisting of L observations from N sensors.}

\item{rate}{rate term in maximum likelihood}

\item{stop_criterion}{if True, prevent overly-long runtimes}

\item{threshold_type}{Which thresholding function to use on the matrix.}

\item{...}{Arguments}
}
\description{
Reconstruction of graphs using maximum likelihood estimation
}
