% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_field.R
\name{mean_field_fit}
\alias{mean_field_fit}
\title{Reconstruction of graphs using the exact mean field}
\usage{
mean_field_fit(
  TS,
  exact = T,
  stop_criterion = T,
  threshold_type = "range",
  ...
)
}
\arguments{
\item{TS}{Array consisting of L observations from N variables.}

\item{exact}{If True, use the exact mean field approximation. If False, use the
naive mean field approximation.}

\item{stop_criterion}{If True, prevent overly-long runtimes. Only applies for exact mean
field.}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{...}{Arguments}
}
\value{
strcuture
}
\description{
Reconstruction of graphs using the exact mean field
}
