% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_information_matrix.R
\name{mutual_information_matrix_fit}
\alias{mutual_information_matrix_fit}
\title{Calculates the mutual information between the probability distributions
of the (binned) values of the time series of pairs of nodes.
First, the mutual information is computed between each pair of
vertices.  Then, a thresholding condition is applied to obtain
edges.
The results dictionary also stores the weight matrix as
`'weights_matrix'` and the thresholded version of the weight matrix
as `'thresholded_matrix'`.}
\usage{
mutual_information_matrix_fit(TS, nbins = 10, threshold_type = "degree", ...)
}
\arguments{
\item{TS}{Matrix consisting of :math:`L` observations from :math:`N` sensors.}

\item{nbins}{number of bins for the pre-processing step (to yield a discrete probability distribution)}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{...}{arguments}
}
\value{
A reconstructed graph with :math:`N` nodes.
}
\description{
Calculates the mutual information between the probability distributions
of the (binned) values of the time series of pairs of nodes.
First, the mutual information is computed between each pair of
vertices.  Then, a thresholding condition is applied to obtain
edges.
The results dictionary also stores the weight matrix as
`'weights_matrix'` and the thresholded version of the weight matrix
as `'thresholded_matrix'`.
}
