% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_transfer_entropy.R
\name{naive_transfer_entropy_fit}
\alias{naive_transfer_entropy_fit}
\title{Graph reconstruction algorithm
Calculates the transfer entropy from i --> j.
The resulting network is asymmetric}
\usage{
naive_transfer_entropy_fit(
  TS,
  delay_max = 1,
  n_bins = 2,
  threshold_type = "range",
  ...
)
}
\arguments{
\item{TS}{array consisting of L observations from N sensors.}

\item{delay_max}{the number of timesteps in the past to aggregate and average in order to get TE_{ij}}

\item{n_bins}{the number of bins to turn values in the time series to categorical
data, which is a pre-processing step to compute entropy.}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{...}{arguments}
}
\value{
a reconstructed graph with N nodes.
}
\description{
Graph reconstruction algorithm
Calculates the transfer entropy from i --> j.
The resulting network is asymmetric
}
