% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thouless_anderson_palmer.R
\name{thouless_anderson_palmer_fit}
\alias{thouless_anderson_palmer_fit}
\title{Infer inter-node coupling weights using a Thouless-Anderson-Palmer mean
field approximation.}
\usage{
thouless_anderson_palmer_fit(TS, threshold_type = "range", ...)
}
\arguments{
\item{TS}{Matrix consisting of L observations from N sensors.}

\item{threshold_type}{Which thresholding function to use on the matrix of weights.}

\item{...}{arguments}
}
\value{
a reconstructed graph.
}
\description{
Infer inter-node coupling weights using a Thouless-Anderson-Palmer mean
field approximation.
}
