% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Utilities for thresholding matrices based on different criteria}
\usage{
threshold(mat, rule, ...)
}
\arguments{
\item{mat}{input matrix}

\item{rule}{A string indicating which thresholding function to invoke.}

\item{...}{Arguments}
}
\description{
Utilities for thresholding matrices based on different criteria
}
